fixes based on recent reviews
