+++ vulcanai2/models/__init__.py
@@ -1,6 +1,5 @@
 from .basenetwork import BaseNetwork
 from .cnn import ConvNet
 from .dnn import DenseNet
-from .utils import *
 
 #__all__ = ['CNN'] #pretty sure this will be useful later
